requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.1.2
flake8==7.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
pycodestyle==2.11.1
pyflakes==3.2.0
sqlparse==0.5.1
Pillow>=11.0.0
djangorestframework>=3.15.2
drf-spectacular>=0.27.2
uwsgi>=2.0.28

